Implement Copy on SmallRef and require Freeze
Requires `copy` to be of a type implementing `Copy`. Chose `[u8; 4]` to
represents 4 arbitrary bytes of data.

Interior mutability breaks the unsafe abstraction. `Freeze` is an
unfortunately private auto trait from libcore. We define our own freeze
but it needs to be understood and implemented manually by consumers of
our library.